requirements.txt
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.1
scipy==1.4.1
tmuxp==1.5.4
torch==1.5.0
tensorflow==2.1.0
tqdm==4.46.0
PyYAML==5.1.2
seaborn==0.9.0
tensorboard==2.4.0
Click==7.0